Refactor folder
