+++ .env.development
@@ -1,0 +1,7 @@
+DB="codehelp"
+DB_PORT="5432"
+DB_PWD="123Choui"
+DB_URL="postgresql://postgres:123Choui@localhost:5432/codehelp"
+DB_USER="postgres"
+PORT="80"
+TOKEN="CodehelpSecret"
